Added Fonts and General Text Styling
Added my 2 fonts from the Google API, added the links into the index.html then utilized the 2 fonts in my style.css. Did all the general text styling of applying the font size and letter spacing for h1 and h2, as well as centering h1 and making the margin bottom 0 for h2. I indented the first line of the first paragraph in each section. Lastly I added the external-link png which will be used later on.